Several enhancement (suppress warnings, fix some errors, enhance portability). In more details: - Increase the warning level for Visual Studio and GCC - Remove several (all) compilation warnings - Replace int by size_t when it is appropriate (sizeof, position) - Remove unused parameters - Keys are now of type char (were of type int, it was wrong) - Replace the non-portable ##__VA_ARGS__ by the portable __VA_ARGS__. As a consequence, new macros (ending with '0') are defined when there are no parameters - Remove some (stupid) syntax errors (';' at the end of some functions). C/C++ is odd: you have to put a ';' when you define a struct, but none when you define a function - A integral type is computed at compile time to store pointers to functions

+++ ADVobfuscator/MetaString1.h
@@ -45,7 +45,7 @@ struct MetaString1
     // Runtime decryption. Most of the time, inlined
     inline const char* decrypt()
     {
-        for(int i = 0; i < sizeof...(I); ++i)
+        for(size_t i = 0; i < sizeof...(I); ++i)
             buffer_[i] = decrypt(buffer_[i]);
         buffer_[sizeof...(I)] = 0;
         return buffer_;
